Fix setParentWindow for NSInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/gnustep_testplant_branch@40219 72102866-910b-0410-8b05-ffd578937521

+++ Source/x11/XGServerWindow.m
@@ -5025,8 +5025,8 @@ _computeDepth(int class, int bpp)
   return NO;
 }
 
-- (void) setParentWindow: (int)parentWin 
-          forChildWindow: (int)childWin
+- (void) setParentWindow: (NSInteger)parentWin
+          forChildWindow: (NSInteger)childWin
 {
   gswindow_device_t	*cwindow;
   gswindow_device_t	*pwindow;
